requirements.txt
streamlit
pandas
python-dotenv
openai
newspaper3k
requests
